# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "novamem",
    "version": "1.1.7",
    "description": "Tiered memory service with hybrid search (keyword + vector + graph), per-user isolation, and project-scoped sub-brains.\n\n**Auth**: every authenticated route accepts either an `nm_…` user bearer (created from the dashboard's API Tokens page) or a Better Auth session cookie/bearer (set by `POST /api/auth/sign-in/email`). Both flow through the same auth hook and resolve to the same user. Better Auth's own routes are mounted at `/api/auth/*` — see the Better Auth documentation for their shapes.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "nm_…",
        "description": "Per-device API token belonging to one user. Carries every right the owning user has — data plane, project CRUD, membership, metrics. Created from the dashboard's API Tokens page; the plaintext is shown once."
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nm.session_token",
        "description": "Better Auth session cookie set by /api/auth/sign-in/email. HttpOnly + SameSite=Lax. Browsers attach it automatically."
      }
    },
...
  },
  "paths": {
...
    "/v1/me/tokens": {
      "get": {
        "summary": "List the caller's API tokens",
        "tags": [
          "tokens"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Mint a new `nm_…` API token for the caller",
        "tags": [
          "tokens"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "maxLength": 128
                  },
                  "scope": {
                    "type": "string",
                    "enum": [
                      "full",
                      "read_only"
                    ]
                  },
                  "project": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "expiresInDays": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 3650
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/tokens/{hash}": {
      "delete": {
        "summary": "Hard-delete a token by its sha256 hash",
        "tags": [
          "tokens"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{64}$"
            },
            "in": "path",
            "name": "hash",
            "required": true
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/projects": {
      "get": {
        "summary": "List projects the caller owns or is a member of",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Create a project (sub-brain)",
        "tags": [
          "projects"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/projects/{id}": {
      "delete": {
        "summary": "Delete a project (owner only)",
        "tags": [
          "projects"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/projects/{id}/members": {
      "get": {
        "summary": "List members of a project (member-only)",
        "tags": [
          "projects"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Invite a user (by exact email) to a project (owner only)",
        "tags": [
          "projects"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "owner",
                      "member"
                    ]
                  }
                },
                "required": [
                  "username"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/projects/{id}/members/{userId}": {
      "delete": {
        "summary": "Remove a member from a project (owner can remove anyone; members can leave)",
        "tags": [
          "projects"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/active-project": {
      "get": {
        "summary": "Get the caller's active project (or null)",
        "tags": [
          "self-service"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "summary": "Set the caller's active project",
        "tags": [
          "self-service"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "project": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  }
                },
                "required": [
                  "project"
                ]
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Clear the caller's active project",
        "tags": [
          "self-service"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/usage": {
      "get": {
        "summary": "The caller's stored-entry count and effective quotas",
        "tags": [
          "self-service"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/changes": {
      "get": {
        "summary": "Per-user memory changelog (created/updated/superseded/deleted/expired)",
        "tags": [
          "self-service"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
            },
            "in": "query",
            "name": "since",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "afterSeq",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 500
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/today": {
      "get": {
        "summary": "Lightweight cross-table activity feed for the user dashboard",
        "tags": [
          "self-service"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/me/onboarding": {
      "get": {
        "summary": "Derived onboarding state for the welcome wizard",
        "tags": [
          "self-service"
        ],
        "security": [
          {
            "SessionCookie": []
          },
          {
            "UserBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "this server"
    }
  ],
...
}